+++ packages/clerk-js/src/ui/components/SignIn/useResetPasswordFactor.tsx
@@ -6,7 +6,7 @@ import { isResetPasswordStrategy } from './utils';
 export function useResetPasswordFactor() {
   const signIn = useCoreSignIn();
 
-  return signIn.supportedFirstFactors.find(({ strategy }) => isResetPasswordStrategy(strategy)) as
+  return signIn.supportedFirstFactors?.find(({ strategy }) => isResetPasswordStrategy(strategy)) as
     | ResetPasswordCodeFactor
     | undefined;
 }