+++ packages/types/src/signIn.ts
@@ -72,7 +72,7 @@ export interface SignInResource extends ClerkResource {
    * @deprecated This attribute will be removed in the next major version
    */
   supportedIdentifiers: SignInIdentifier[];
-  supportedFirstFactors: SignInFirstFactor[];
+  supportedFirstFactors: SignInFirstFactor[] | null;
   supportedSecondFactors: SignInSecondFactor[] | null;
   firstFactorVerification: VerificationResource;
   secondFactorVerification: VerificationResource;