+++ packages/tanstack-start/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @clerk/tanstack-start
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [[`7e0ced3da`](https://github.com/clerk/javascript/commit/7e0ced3da94f41056bc4445d163d3b615afb6ab1), [`b6f0613dc`](https://github.com/clerk/javascript/commit/b6f0613dc9d8b0bab41cfabbaa8621b126e3bdf5)]:
+  - @clerk/shared@2.5.1
+  - @clerk/clerk-react@5.4.1
+  - @clerk/types@4.13.1
+  - @clerk/backend@1.6.3
+
 ## 0.2.0
 
 ### Minor Changes